🐝🤖 make Vite configs native-loader compatible
Vite plans to make native config loading the default, but this config
graph relied on bundled transforms for CommonJS-classified TypeScript,
TypeScript enums, emitted .js import resolution, and __dirname.

Move the config dependencies and Vitest config to explicit ESM, preserve
compatibility wrappers for application imports, and rewrite emitted
extensions so native source loading and compiled output both work.
