+++ vite.config-archive.mts
@@ -1,4 +1,4 @@
-import { ViteEntryPoint } from "./site/viteConstants.ts"
+import { ViteEntryPoint } from "./site/viteConstants.mts"
 import { defineViteConfigForEntrypoint } from "./vite.config-common.mts"
 
 export default defineViteConfigForEntrypoint(ViteEntryPoint.Archive)
